delete完成/ファイル整理, README編集
